+++ src/MVector.hpp
@@ -4,8 +4,8 @@
 #include <cmath>
 
 struct MVector {
-    double x = 0.0;
-    double y = 0.0;
+    float x = 0.0;
+    float y = 0.0;
 
     MVector& operator+=(const MVector& rhs) {
         this->x += rhs.x;
@@ -18,7 +18,7 @@ struct MVector {
         return lhs;
     }
 
-    friend MVector operator*(MVector lhs, const double& rhs) {
+    friend MVector operator*(MVector lhs, const float& rhs) {
         lhs.x *= rhs;
         lhs.y *= rhs;
         return lhs;
@@ -36,14 +36,14 @@ struct MVector {
     }
 
     MVector direction() {
-        double m = this->magnitude();
+        float m = this->magnitude();
         MVector v;
         v.x = this->x / m;
         v.y = this->y / m;
         return v;
     }
 
-    double magnitude() {
+    float magnitude() {
         return sqrt(pow(x, 2) + pow(y, 2));
     }
 };
